requirements.txt
# Core trading dependencies
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
websockets>=11.0.0
aiohttp>=3.8.0

# Technical analysis
TA-Lib>=0.4.25
scipy>=1.10.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0
PyMySQL>=1.1.0

# Configuration and environment
PyYAML>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Scheduling
APScheduler>=3.10.0

# Web framework (for dashboard)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Notifications
python-telegram-bot>=20.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security
cryptography>=41.0.0

# Utilities
click>=8.1.0
requests>=2.31.0
httpx>=0.24.0

# Data processing
joblib>=1.3.0
